requirements.txt
Flask==3.1.0
beautifulsoup4==4.12.3
requests==2.32.3
google-generativeai==0.8.6
python-dotenv==1.2.1
